+++ resharper/resharper-unity/src/Settings/LangVersionSetting.cs
@@ -8,6 +8,8 @@ using JetBrains.Lifetimes;
 using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.Properties.CSharp;
 using JetBrains.ReSharper.Plugins.Unity.ProjectModel.Caches;
+using JetBrains.ReSharper.Psi;
+using JetBrains.ReSharper.Psi.CSharp;
 using JetBrains.ReSharper.Psi.CSharp.Impl;
 using JetBrains.Util;
 
@@ -19,14 +21,17 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
         private readonly ISettingsSchema mySettingsSchema;
         private readonly ILogger myLogger;
         private readonly UnityProjectFileCacheProvider myUnityProjectFileCache;
+        private readonly ILanguageLevelProjectProperty<CSharpLanguageLevel, CSharpLanguageVersion> myLanguageLevelProjectProperty;
         private static readonly Version ourVersion46 = new Version(4, 6);
 
         public LangVersionSetting(ISettingsSchema settingsSchema, ILogger logger,
-                                  UnityProjectFileCacheProvider unityProjectFileCache)
+                                  UnityProjectFileCacheProvider unityProjectFileCache,
+                                  ILanguageLevelProjectProperty<CSharpLanguageLevel, CSharpLanguageVersion> languageLevelProjectProperty)
         {
             mySettingsSchema = settingsSchema;
             myLogger = logger;
             myUnityProjectFileCache = unityProjectFileCache;
+            myLanguageLevelProjectProperty = languageLevelProjectProperty;
         }
 
         public void InitialiseProjectSettings(Lifetime projectLifetime, IProject project,
@@ -52,14 +57,14 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
             // * Unity prior to 5.5 uses an old mono compiler that only supports C# 4
             // * Unity 5.5 and later adds C# 6 support as an option. This is enabled by setting
             //   the API compatibility level to NET_4_6
-            // * The CSharp60Support plugin replaces the compiler with either C# 6 or C# 7.0
-            //   It can be recognised by a folder called `CSharp60Support` or `CSharp70Support`
+            // * The CSharp60Support plugin replaces the compiler with either C# 6 or C# 7.0 or 8.0
+            //   It can be recognised by a folder called `CSharp60Support` or `CSharp70Support` or `CSharp80Support`
             //   in the root of the project
             //   (https://bitbucket.org/alexzzzz/unity-c-5.0-and-6.0-integration)
-            // * Note that since Unity 2017.2, we've been special-cased in the Unity csproj generation
+            // * Note that since Unity 2017.2 till 2018.3, we've been special-cased in the Unity csproj generation
             //   and we've been getting v4.5 for old runtime and default values (4.7.1) for new. So where
             //   it says 3.5 below, that depends on the version of Unity. Older versions will give us 3.5,
-            //   newer versions 4.5. We don't need this special case, so hopefully it will be removed in 2018.3
+            //   newer versions 4.5.
             //
             // Scenarios:
             // * No VSTU installed (including Unity 5.5)
@@ -68,7 +73,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
             //   .csproj has NO `LangVersion`. `TargetFrameworkVersion` will be `v3.5`
             // * Later versions of VSTU
             //   `LangVersion` is correctly set to "4". `TargetFrameworkVersion` will be `v3.5`
-            //   OR `LangVersion` is set to "6" or "latest". `TargetFrameworkVersion` will be `v4.7.1`
+            //   OR `LangVersion` is set to "6" or "latest".
             // * VSTU for 5.5
             //   `LangVersion` is set to "default". `TargetFrameworkVersion` will be `v3.5` or `v4.6`
             //   Note that "default" for VS"15" or Rider will be C# 7.0!
@@ -79,11 +84,14 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
             //   .csproj has NO `LangVersion`
             //   `TargetFrameworkVersion` is NOT accurate (support for C# 6 is not dependent on/trigger by .net 4.6)
             //   Look for `CSharp60Support` or `CSharp70Support` folders
+            // * Unity 2018.x+. LangVersion is `latest`. MSBuild 16 treats `newest` as C# 8. Re# and Rider start suggesting it.
             //
             // Actions:
-            // If `LangVersion` is missing or "default"
+            // * If `LangVersion` is missing or "default"
             // then override based on `TargetFrameworkVersion` or presence of `CSharp60Support`/`CSharp70Support`
             // else do nothing
+            // * If `LangVersion` is "latest"
+            // then override based on `CSharp80Support` presence or LangVersion matching Roslyn bundled in Unity
             //
             // Notes:
             // * Unity and VSTU have two separate .csproj routines. VSTU adds extra references,
@@ -95,8 +103,6 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
             // * `LangVersion` can be conditionally specified, which makes checking for "default" awkward
             // * If Unity3dRider + CSharp60Support are both installed, last write wins
             //   Order of post-processing is non-deterministic, so Rider's LangVersion might be removed
-            // * Unity3dRider can set `TargetFrameworkVersion` to `v4.5` on non-Windows machines to fix
-            //   an issue resolving System.Linq
             #endregion
 
             var languageLevel = ReSharperSettingsCSharpLanguageLevel.Default;
@@ -113,6 +119,21 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
                     languageLevel = IsTargetFrameworkAtLeast46(project)
                         ? ReSharperSettingsCSharpLanguageLevel.CSharp60
                         : ReSharperSettingsCSharpLanguageLevel.CSharp40;
+                }
+            }
+
+            if (IsLangVersionLatest(project))
+            {
+                if (project.Location.CombineWithShortName("CSharp80Support").ExistsDirectory) // https://forum.unity.com/threads/would-the-roslyn-compiler-compile-c-8-0-preview.598069/
+                    languageLevel = ReSharperSettingsCSharpLanguageLevel.CSharp80;
+                else
+                {
+                    var appPath = myUnityProjectFileCache.GetAppPath(project);
+                    var contentPath = UnityInstallationFinder.GetApplicationContentsPath(appPath);
+                    var roslynDir = contentPath.Combine(@"Tools\Roslyn");
+
+                    if (roslynDir.ExistsDirectory)
+                        languageLevel = myLanguageLevelProjectProperty.GetLatestAvailableLanguageLevel(roslynDir).ToSettingsLanguageLevel();   
                 }
             }
 
@@ -140,6 +161,19 @@ namespace JetBrains.ReSharper.Plugins.Unity.Settings
                 }
             }
             return true;
+        }
+        
+        private bool IsLangVersionLatest(IProject project)
+        {
+            foreach (var configuration in project.ProjectProperties.ActiveConfigurations.Configurations)
+            {
+                if (configuration is ICSharpProjectConfiguration csharpConfiguration)
+                {
+                    if (csharpConfiguration.LanguageVersion == CSharpLanguageVersion.Latest)
+                        return true;
+                }
+            }
+            return false;
         }
 
         private bool IsTargetFrameworkAtLeast46(IProject project)